Split the generator into a src/ package
One 1,551-line module becomes nine, under src/translator_diagram/, so that
people working on different parts of the tool stop sharing one file:

  model.py       Component, index_by_id
  naming.py      SVG ids, clone ids, output filename stems
  colors.py      ColorAssigner, owner colours, the palette constants
  loading.py     cell parsers, load_components, the Google Sheet download
  validation.py  validate
  render.py      build_graph and the per-layer sub-figures
  legend.py      the owner and edge-style legends
  export.py      write_json
  cli.py         the click command and the run sequence

The modules are the subject sections the test file already uses, so the tests
split onto them one-for-one in the next commit.

Two things are deliberate. build_graph and build_layer_subgraph stay in one
module: they duplicate the edge-suppression rules, drift between them has
caused a real bug, and a file boundary would make the next drift easier to
miss. And the colour constants live in colors.py rather than render.py so that
render and legend can share them without importing each other — imports run
model -> naming -> {validation, export, render}, colors -> {render, legend},
render -> legend, with cli on top and nothing importing cli.

The Google Sheet download moves out of main() into loading.download_sheet_csv,
which is the only code that moved rather than being carried across.

owner-colors.csv moves to src/translator_diagram/data/ so an installed copy
ships with colours; the next commits give it a resolution order and a flag.
Packaging follows: the entry point is now translator_diagram.cli:main and the
wheel is built from the package directory.

Verified by generating from a sample CSV before and after the split: the .dot,
.json, .svg and every .png are byte-identical, and the .pdf differs only in
its embedded CreationDate.
