+++ pyproject.toml
@@ -10,7 +10,7 @@ dependencies = [
 ]
 
 [project.scripts]
-generate-diagram = "generate_diagram:main"
+generate-diagram = "translator_diagram.cli:main"
 
 [dependency-groups]
 dev = [
@@ -22,11 +22,10 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-# The module and its owner-colors.csv must land side by side in the wheel:
-# DEFAULT_OWNER_COLORS_PATH resolves the CSV relative to __file__. Listing
-# them explicitly also keeps .git/, data/, tests/ and .venv/ out of the build,
-# which "packages = [.]" would sweep in now that the layout is flat.
-only-include = ["generate_diagram.py", "owner-colors.csv"]
+# The package carries its own copy of owner-colors.csv in data/, so an
+# installed copy has colours without a checkout to read them from. Naming the
+# package directory keeps .git/, data/, tests/ and .venv/ out of the build.
+packages = ["src/translator_diagram"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
