+++ src/translator_diagram/__init__.py
@@ -1,0 +1,1 @@
+"""Generate dependency diagrams for Translator platform components."""
