+++ src/translator_diagram/data/owner-colors.csv
